* rework the event subscriber
