+++ Bootstrap.php
@@ -194,9 +194,6 @@ class Shopware_Plugins_Backend_ExitBBlisstribute_Bootstrap extends Shopware_Comp
 
         $em = $this->get('models');
         if (version_compare($version, '0.2.2', '<=')) {
-            $this->subscribeEvent('Shopware\Models\Property\Group::prePersist', 'prePersistProperty');
-            $this->subscribeEvent('Shopware\Models\Property\Group::preRemove', 'preRemoveProperty');
-
             $classMetadata = $em->getClassMetadata('Shopware\CustomModels\Blisstribute\BlisstributeArticleType');
             $this->handleTableInstall($classMetadata);
 
@@ -215,50 +212,7 @@ class Shopware_Plugins_Backend_ExitBBlisstribute_Bootstrap extends Shopware_Comp
             $this->get('db')->query($sql);
         }
 
-        if (version_compare($version, '0.2.10', '<=')) {
-            $this->subscribeEvent('Shopware\Models\Article\Article::prePersist', 'prePersistArticle');
-        }
-
-        if (version_compare($version, '0.2.25', '<=')) {
-            $this->subscribeEvent('Shopware\Models\Order\Order::prePersist', 'prePersistOrder');
-        }
-
-        if (version_compare($version, '0.2.29', '<=')) {
-            $this->subscribeEvent(
-                'Enlight_Controller_Dispatcher_ControllerPath_Backend_BlisstributeArticleSyncCron',
-                'getArticleSyncCronController'
-            );
-        }
-
-        if (version_compare($version, '0.2.30', '<=')) {
-            $this->subscribeEvent(
-                'Enlight_Controller_Dispatcher_ControllerPath_Backend_BlisstributeOrderSyncCron',
-                'getOrderSyncCronController'
-            );
-        }
-
         if (version_compare($version, '0.3.0', '<')) {
-            $this->subscribeEvent(
-                'Enlight_Controller_Dispatcher_ControllerPath_Backend_BlisstributeShopMapping',
-                'getShopMappingController'
-            );
-
-            $this->subscribeEvent(
-                'Enlight_Controller_Dispatcher_ControllerPath_Backend_BlisstributeCouponMapping',
-                'getCouponMappingController'
-            );
-
-            $this->subscribeEvent(
-                'Enlight_Controller_Dispatcher_ControllerPath_Api_Btarticlestock',
-                'onGetBtarticlestockApiController'
-            );
-
-            $this->subscribeEvent('Shopware\Models\Shop\Shop::prePersist', 'prePersistShop');
-            $this->subscribeEvent('Shopware\Models\Shop\Shop::preRemove', 'preRemoveShop');
-
-            $this->subscribeEvent('Shopware\Models\Voucher\Voucher::prePersist', 'prePersistVoucher');
-            $this->subscribeEvent('Shopware\Models\Voucher\Voucher::preRemove', 'preRemoveVoucher');
-
             $classMetadata = $em->getClassMetadata('Shopware\CustomModels\Blisstribute\BlisstributeShop');
             $this->handleTableInstall($classMetadata);
 
@@ -308,17 +262,6 @@ class Shopware_Plugins_Backend_ExitBBlisstribute_Bootstrap extends Shopware_Comp
             );
         }
 
-        if (version_compare($version, '0.3.7', '<')) {
-            $this->subscribeEvent(
-                'Shopware\Models\Attribute\Voucher::prePersist',
-                'prePersistVoucherAttribute'
-            );
-        }
-
-        if (version_compare($version, '0.3.11', '<')) {
-            $this->subscribeEvent('Shopware\Models\Order\Order::preRemove', 'preRemoveOrder');
-        }
-
         if (version_compare($version, '0.3.15', '<')) {
             $paymentMappings = $this->get('db')->fetchAll("SELECT * FROM s_plugin_blisstribute_payment WHERE mapping_class_name != ''");
 
@@ -335,16 +278,7 @@ class Shopware_Plugins_Backend_ExitBBlisstribute_Bootstrap extends Shopware_Comp
         }
         
         if (version_compare($version, '0.6.3', '<')) {
-            // check if plugin ExitBBlisstribute is installed
-			$plugin = Shopware()->Models()->getRepository('Shopware\Models\Plugin\Plugin')->findOneBy([
-				'name' => 'ExitBBlisstribute',
-			]);
-			
-			if($plugin)
-			{
-				Shopware()->Db()->query("DELETE FROM s_core_subscribes WHERE pluginID = ?", [$plugin->getId()]);
-                $this->subscribeEvents();
-			}
+            array('success' => false, 'message' => 'Bitte das Plugin neu installieren.');
         }
 
         return array('success' => true, 'invalidateCache' => array('backend', 'proxy', 'config', 'frontend'));
@@ -506,125 +440,6 @@ class Shopware_Plugins_Backend_ExitBBlisstribute_Bootstrap extends Shopware_Comp
         );
 
         $this->subscribeEvent(
-            'Enlight_Controller_Dispatcher_ControllerPath_Backend_BlisstributeArticle',
-            'getArticleSyncController'
-        );
-
-        $this->subscribeEvent(
-            'Enlight_Controller_Dispatcher_ControllerPath_Backend_BlisstributeArticleSyncCron',
-            'getArticleSyncCronController'
-        );
-
-        $this->subscribeEvent(
-            'Enlight_Controller_Dispatcher_ControllerPath_Backend_BlisstributeArticleType',
-            'getArticleTypeController'
-        );
-
-        $this->subscribeEvent(
-            'Enlight_Controller_Dispatcher_ControllerPath_Backend_BlisstributeOrder',
-            'getOrderController'
-        );
-
-        $this->subscribeEvent(
-            'Enlight_Controller_Dispatcher_ControllerPath_Backend_BlisstributeOrderSyncCron',
-            'getOrderSyncCronController'
-        );
-
-        $this->subscribeEvent(
-            'Enlight_Controller_Dispatcher_ControllerPath_Backend_BlisstributeShipmentMapping',
-            'getShipmentMappingController'
-        );
-
-        $this->subscribeEvent(
-            'Enlight_Controller_Dispatcher_ControllerPath_Backend_BlisstributePaymentMapping',
-            'getPaymentMappingController'
-        );
-
-        $this->subscribeEvent(
-            'Enlight_Controller_Dispatcher_ControllerPath_Backend_BlisstributeShopMapping',
-            'getShopMappingController'
-        );
-
-        $this->subscribeEvent(
-            'Enlight_Controller_Dispatcher_ControllerPath_Backend_BlisstributeCouponMapping',
-            'getCouponMappingController'
-        );
-
-        $this->subscribeEvent('Shopware_Modules_Order_SaveOrder_ProcessDetails', 'onRegisterOrder');
-        $this->subscribeEvent('Shopware_Modules_Order_SendMail_BeforeSend', 'onOrderFinished');
-
-        // model events
-        $this->subscribeEvent('Shopware\Models\Order\Order::prePersist', 'prePersistOrder');
-        $this->subscribeEvent('Shopware\Models\Order\Order::preRemove', 'preRemoveOrder');
-
-        $this->subscribeEvent('Shopware\Models\Article\Article::prePersist', 'prePersistArticle');
-        $this->subscribeEvent('Shopware\Models\Article\Article::preUpdate', 'preUpdateArticle');
-        $this->subscribeEvent('Shopware\Models\Article\Article::preRemove', 'preRemoveArticle');
-
-        $this->subscribeEvent('Shopware\Models\Article\Detail::prePersist', 'prePersistDetail');
-        $this->subscribeEvent('Shopware\Models\Article\Detail::preUpdate', 'preUpdateDetail');
-        $this->subscribeEvent('Shopware\Models\Article\Detail::preRemove', 'preRemoveDetail');
-
-        $this->subscribeEvent('Shopware\Models\Property\Group::prePersist', 'prePersistProperty');
-        $this->subscribeEvent('Shopware\Models\Property\Group::preRemove', 'preRemoveProperty');
-
-        $this->subscribeEvent('Shopware\Models\Shop\Shop::prePersist', 'prePersistShop');
-        $this->subscribeEvent('Shopware\Models\Shop\Shop::preRemove', 'preRemoveShop');
-
-        $this->subscribeEvent('Shopware\Models\Voucher\Voucher::prePersist', 'prePersistVoucher');
-        $this->subscribeEvent('Shopware\Models\Voucher\Voucher::preRemove', 'preRemoveVoucher');
-
-        // blisstribute model events
-        $this->subscribeEvent(
-            'Shopware\CustomModels\Blisstribute\BlisstributeOrder::prePersist',
-            'prePersistBlisstributeOrder'
-        );
-
-        $this->subscribeEvent(
-            'Shopware\CustomModels\Blisstribute\BlisstributeOrder::preUpdate',
-            'preUpdateBlisstributeOrder'
-        );
-
-        $this->subscribeEvent(
-            'Shopware\CustomModels\Blisstribute\BlisstributeShippingRequest::prePersist',
-            'prePersistBlisstributeShippingRequest'
-        );
-
-        $this->subscribeEvent(
-            'Shopware\CustomModels\Blisstribute\BlisstributeShippingRequest::preUpdate',
-            'preUpdateBlisstributeShippingRequest'
-        );
-        $this->subscribeEvent(
-            'Shopware\CustomModels\Blisstribute\BlisstributeShippingRequestItems::prePersist',
-            'prePersistBlisstributeShippingRequestItem'
-        );
-
-        $this->subscribeEvent(
-            'Shopware\CustomModels\Blisstribute\BlisstributeShippingRequestItems::preUpdate',
-            'preUpdateBlisstributeShippingRequestItem'
-        );
-
-        $this->subscribeEvent(
-            'Shopware\CustomModels\Blisstribute\BlisstributeArticleType::prePersist',
-            'prePersistBlisstributeArticleType'
-        );
-
-        $this->subscribeEvent(
-            'Shopware\CustomModels\Blisstribute\BlisstributeArticleType::preUpdate',
-            'preUpdateBlisstributeArticleType'
-        );
-
-        $this->subscribeEvent(
-            'Shopware\CustomModels\Blisstribute\BlisstributeArticle::prePersist',
-            'prePersistBlisstributeArticle'
-        );
-
-        $this->subscribeEvent(
-            'Shopware\CustomModels\Blisstribute\BlisstributeArticle::preUpdate',
-            'preUpdateBlisstributeArticle'
-        );
-
-        $this->subscribeEvent(
             'Enlight_Controller_Action_PostDispatch_Backend_Article',
             'postDispatchBackendArticle'
         );
@@ -654,11 +469,6 @@ class Shopware_Plugins_Backend_ExitBBlisstribute_Bootstrap extends Shopware_Comp
             'onActionPostDispatchSecureBackendIndex'
         );
 
-        $this->subscribeEvent(
-            'Shopware\Models\Attribute\Voucher::prePersist',
-            'prePersistVoucherAttribute'
-        );
-
         $this->subscribeEvent('Enlight_Controller_Front_StartDispatch', 'startDispatch');
 
         $this->subscribeEvent('Enlight_Bootstrap_InitResource_blisstribute.google_address_validator', 'onGetGoogleAddressValidator');
@@ -679,18 +489,14 @@ class Shopware_Plugins_Backend_ExitBBlisstribute_Bootstrap extends Shopware_Comp
         $this->registerSnippets();
 
         $subscribers = array(
-            new CronSubscriber(
-                Shopware()->Container(),
-                $this->Path(),
-                $this->Application(),
-                $this->get('db')
-            )
+            new CronSubscriber(),
+            new ControllerPathSubscriber(),
+            new ModelSubscriber()
         );
 
         foreach ($subscribers as $subscriber) {
             $this->get('events')->addSubscriber($subscriber);
         }
-
     }
 
     /**
@@ -792,829 +598,6 @@ class Shopware_Plugins_Backend_ExitBBlisstribute_Bootstrap extends Shopware_Comp
         if ($arguments->getRequest()->getActionName() === 'load') {
             $view->extendsTemplate('backend/index/view/exitb_blisstribute/menu.js');
         }
-    }
-
-    /**
-     * return article type controller path
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return string
-     */
-    public function getArticleTypeController(Enlight_Event_EventArgs $eventArgs)
-    {
-        $this->registerTemplateDir();
-        $this->Application()->Snippets()->addConfigDir(__DIR__ . '/Snippets/');
-        return $this->Path() . 'Controllers/Backend/BlisstributeArticleType.php';
-    }
-
-    /**
-     * return article sync controller path
-
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return string
-     */
-    public function getArticleSyncController(Enlight_Event_EventArgs $eventArgs)
-    {
-        $this->registerTemplateDir();
-        $this->Application()->Snippets()->addConfigDir(__DIR__ . '/Snippets/');
-        return $this->Path() . 'Controllers/Backend/BlisstributeArticle.php';
-    }
-
-    /**
-     * return article sync cron controller path
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return string
-     */
-    public function getArticleSyncCronController(Enlight_Event_EventArgs $eventArgs)
-    {
-        $this->registerTemplateDir();
-        $this->Application()->Snippets()->addConfigDir(__DIR__ . '/Snippets/');
-        return $this->Path() . 'Controllers/Backend/BlisstributeArticleSyncCron.php';
-    }
-
-    /**
-     * return order export controller path
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return string
-     */
-    public function getOrderController(Enlight_Event_EventArgs $eventArgs)
-    {
-        $this->registerTemplateDir();
-        $this->Application()->Snippets()->addConfigDir(__DIR__ . '/Snippets/');
-        return $this->Path() . 'Controllers/Backend/BlisstributeOrder.php';
-    }
-
-    /**
-     * return order sync cron controller path
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return string
-     */
-    public function getOrderSyncCronController(Enlight_Event_EventArgs $eventArgs)
-    {
-        $this->registerTemplateDir();
-        $this->Application()->Snippets()->addConfigDir(__DIR__ . '/Snippets/');
-        return $this->Path() . 'Controllers/Backend/BlisstributeOrderSyncCron.php';
-    }
-
-    /**
-     * return shipment mapping controller path
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return string
-     */
-    public function getShipmentMappingController(Enlight_Event_EventArgs $eventArgs)
-    {
-        $this->registerTemplateDir();
-        $this->Application()->Snippets()->addConfigDir(__DIR__ . '/Snippets/');
-        return $this->Path() . 'Controllers/Backend/BlisstributeShipmentMapping.php';
-    }
-
-    /**
-     * return payment mapping controller path
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return string
-     */
-    public function getPaymentMappingController(Enlight_Event_EventArgs $eventArgs)
-    {
-        $this->registerTemplateDir();
-        $this->Application()->Snippets()->addConfigDir(__DIR__ . '/Snippets/');
-        return $this->Path() . 'Controllers/Backend/BlisstributePaymentMapping.php';
-    }
-
-    /**
-     * return payment mapping controller path
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return string
-     */
-    public function getShopMappingController(Enlight_Event_EventArgs $eventArgs)
-    {
-        $this->registerTemplateDir();
-        $this->Application()->Snippets()->addConfigDir(__DIR__ . '/Snippets/');
-        return $this->Path() . 'Controllers/Backend/BlisstributeShopMapping.php';
-    }
-
-    /**
-     * return payment mapping controller path
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return string
-     */
-    public function getCouponMappingController(Enlight_Event_EventArgs $eventArgs)
-    {
-        $this->registerTemplateDir();
-        $this->Application()->Snippets()->addConfigDir(__DIR__ . '/Snippets/');
-        return $this->Path() . 'Controllers/Backend/BlisstributeCouponMapping.php';
-    }
-
-    /**
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return bool
-     */
-    public function prePersistOrder(Enlight_Event_EventArgs $eventArgs)
-    {
-        $this->registerOrder($eventArgs);
-        return true;
-    }
-
-    /**
-     * @param Enlight_Event_EventArgs $args
-     *
-     * @return void
-     */
-    public function prePersistArticle(Enlight_Event_EventArgs $args)
-    {
-        $modelManager = $this->get('models');
-
-        /** @var \Shopware\Models\Article\Article $article */
-        $article = $args->get('entity');
-
-        $blisstributeArticle = new Shopware\CustomModels\Blisstribute\BlisstributeArticle();
-        $blisstributeArticle->setLastCronAt(new DateTime())
-            ->setArticle($article)
-            ->setTriggerSync(true)
-            ->setTries(0)
-            ->setComment(null);
-
-        $modelManager->persist($blisstributeArticle);
-        $modelManager->flush();
-    }
-
-    /**
-     * @param Enlight_Event_EventArgs $args
-     *
-     * @return void
-     */
-    public function preUpdateArticle(Enlight_Event_EventArgs $args)
-    {
-        $modelManager = $this->get('models');
-
-        /** @var \Shopware\Models\Article\Article $article */
-        $article = $args->get('entity');
-
-        $repository = $modelManager->getRepository('Shopware\CustomModels\Blisstribute\BlisstributeArticle');
-        $blisstributeArticle = $repository->findOneBy(array('article' => $article));
-        if ($blisstributeArticle === null) {
-            $blisstributeArticle = new Shopware\CustomModels\Blisstribute\BlisstributeArticle();
-            $blisstributeArticle->setArticle($article);
-        }
-        
-        if ($blisstributeArticle->isTriggerSync()) {
-            return;
-        }
-
-        $blisstributeArticle->setLastCronAt(new DateTime())
-            ->setTriggerSync(true)
-            ->setTries(0)
-            ->setComment(null);
-
-        $modelManager->persist($blisstributeArticle);
-        $modelManager->flush();
-    }
-
-    /**
-     * @param Enlight_Event_EventArgs $args
-     *
-     * @return void
-     */
-    public function preRemoveArticle(Enlight_Event_EventArgs $args)
-    {
-        $modelManager = $this->get('models');
-
-        /** @var \Shopware\Models\Article\Article $article */
-        $article = $args->get('entity');
-
-        $repository = $modelManager->getRepository('Shopware\CustomModels\Blisstribute\BlisstributeArticle');
-        $blisstributeArticle = $repository->findOneBy(array('article' => $article));
-        if ($blisstributeArticle === null) {
-            return;
-        }
-
-        $modelManager->remove($blisstributeArticle);
-        $modelManager->flush();
-    }
-
-    /**
-     * @param Enlight_Event_EventArgs $args
-     *
-     * @return void
-     */
-    public function prePersistDetail(Enlight_Event_EventArgs $args)
-    {
-        $modelManager = $this->get('models');
-        
-        /** @var \Shopware\Models\Article\Detail $detail */
-        $detail = $args->get('entity');
-
-        // load article
-        $repository = $modelManager->getRepository('Shopware\CustomModels\Blisstribute\BlisstributeArticle');
-        $blisstributeArticle = $repository->findOneBy(array('article' => $detail->getArticle()));
-        if ($blisstributeArticle === null) {
-            $blisstributeArticle = new Shopware\CustomModels\Blisstribute\BlisstributeArticle();
-            $blisstributeArticle->setArticle($detail->getArticle());
-        }
-        
-        if ($blisstributeArticle->isTriggerSync()) {
-            return;
-        }
-
-        $blisstributeArticle->setLastCronAt(new DateTime())
-            ->setTriggerSync(true)
-            ->setTries(0)
-            ->setComment(null);
-
-        $modelManager->persist($blisstributeArticle);
-        $modelManager->flush();
-    }
-
-    /**
-     * @param Enlight_Event_EventArgs $args
-     *
-     * @return void
-     */
-    public function preUpdateDetail(Enlight_Event_EventArgs $args)
-    {
-        $modelManager = $this->get('models');
-
-        /** @var \Shopware\Models\Article\Detail $detail */
-        $detail = $args->get('entity');
-
-        $articleRepository = $modelManager->getRepository('Shopware\CustomModels\Blisstribute\BlisstributeArticle');
-
-        /* @var Shopware\CustomModels\Blisstribute\BlisstributeArticle $article */
-        $blisstributeArticle = $articleRepository->findOneBy(array('article' => $detail->getArticle()));
-
-        if ($blisstributeArticle === null) {
-            $blisstributeArticle = new Shopware\CustomModels\Blisstribute\BlisstributeArticle();
-            $blisstributeArticle->setArticle($detail->getArticle());
-        }
-        
-        if ($blisstributeArticle->isTriggerSync()) {
-            return;
-        }
-
-        $blisstributeArticle->setLastCronAt(new DateTime())
-            ->setTriggerSync(true)
-            ->setTries(0)
-            ->setComment(null);
-
-        $modelManager->persist($blisstributeArticle);
-        $modelManager->flush();
-    }
-
-    /**
-     * @param Enlight_Event_EventArgs $args
-     *
-     * @return void
-     */
-    public function preRemoveDetail(Enlight_Event_EventArgs $args)
-    {
-        $modelManager = $this->get('models');
-
-        /** @var \Shopware\Models\Article\Detail $detail */
-        $detail = $args->get('entity');
-
-        $repository = $modelManager->getRepository('Shopware\CustomModels\Blisstribute\BlisstributeArticle');
-
-        /* @var Shopware\CustomModels\Blisstribute\BlisstributeArticle $blisstributeArticle */
-        $blisstributeArticle = $repository->findOneBy(array('article' => $detail->getArticle()));
-
-        if ($blisstributeArticle === null or $blisstributeArticle->isDeleted()) {
-            return;
-        }
-
-        $blisstributeArticle->setTriggerSync(true)
-            ->setTries(0)
-            ->setComment(null)
-            ->setDeleted(true);
-
-        $modelManager->persist($blisstributeArticle);
-        $modelManager->flush();
-    }
-
-    /**
-     * @param Enlight_Event_EventArgs $args
-     *
-     * @return void
-     */
-    public function prePersistProperty(Enlight_Event_EventArgs $args)
-    {
-        $modelManager = $this->get('models');
-        $entity = $args->get('entity');
-
-        $articleType = new Shopware\CustomModels\Blisstribute\BlisstributeArticleType();
-        $articleType->setFilter($entity);
-        $articleType->setArticleType(0);
-
-        $modelManager->persist($articleType);
-        $modelManager->flush();
-    }
-
-    /**
-     * @param Enlight_Event_EventArgs $args
-     *
-     * @return void
-     */
-    public function preRemoveProperty(Enlight_Event_EventArgs $args)
-    {
-        $modelManager = $this->get('models');
-        $entity = $args->get('entity');
-
-        /* @var Shopware\CustomModels\Blisstribute\BlisstributeArticleTypeRepository $articleTypeRepository */
-        $articleTypeRepository = $modelManager->getRepository('Shopware\CustomModels\Blisstribute\BlisstributeArticleType');
-        $articleType = $articleTypeRepository->fetchByFilterType($entity->getId());
-        if ($articleType === null) {
-            return;
-        }
-
-        $modelManager->remove($articleType);
-        $modelManager->flush();
-    }
-
-    /**
-     * blisstribute order event fired before db insert
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return void
-     */
-    public function prePersistBlisstributeOrder(Enlight_Event_EventArgs $eventArgs)
-    {
-        $currentTime = new DateTime();
-
-        /** @var \Shopware\CustomModels\Blisstribute\BlisstributeOrder $entity */
-        $entity = $eventArgs->get('entity');
-        $entity->setCreatedAt($currentTime)
-            ->setModifiedAt($currentTime);
-    }
-
-    /**
-     * blisstribute event fired before db update
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return void
-     */
-    public function preUpdateBlisstributeOrder(Enlight_Event_EventArgs $eventArgs)
-    {
-        $entity = $eventArgs->get('entity');
-        $entity->setModifiedAt(new DateTime());
-    }
-
-    /**
-     * blisstribute order event fired before db insert
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return void
-     */
-    public function prePersistBlisstributeShippingRequest(Enlight_Event_EventArgs $eventArgs)
-    {
-        $currentTime = new DateTime();
-
-        /** @var \Shopware\CustomModels\Blisstribute\BlisstributeShippingRequest $entity */
-        $entity = $eventArgs->get('entity');
-        $entity->setCreatedAt($currentTime)
-            ->setModifiedAt($currentTime);
-    }
-
-    /**
-     * blisstribute event fired before db update
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return void
-     */
-    public function preUpdateBlisstributeShippingRequest(Enlight_Event_EventArgs $eventArgs)
-    {
-        /** @var \Shopware\CustomModels\Blisstribute\BlisstributeShippingRequest $entity */
-        $entity = $eventArgs->get('entity');
-        $entity->setModifiedAt(new DateTime());
-    }
-
-    /**
-     * blisstribute order event fired before db insert
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return void
-     */
-    public function prePersistBlisstributeShippingRequestItem(Enlight_Event_EventArgs $eventArgs)
-    {
-        $currentTime = new DateTime();
-
-        /** @var \Shopware\CustomModels\Blisstribute\BlisstributeShippingRequestItems $entity */
-        $entity = $eventArgs->get('entity');
-        $entity->setCreatedAt($currentTime)
-            ->setModifiedAt($currentTime);
-    }
-
-    /**
-     * blisstribute event fired before db update
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return void
-     */
-    public function preUpdateBlisstributeShippingRequestItem(Enlight_Event_EventArgs $eventArgs)
-    {
-        /** @var \Shopware\CustomModels\Blisstribute\BlisstributeShippingRequestItems $entity */
-        $entity = $eventArgs->get('entity');
-        $entity->setModifiedAt(new DateTime());
-    }
-
-    /**
-     * article type event fired before create entity
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return void
-     */
-    public function prePersistBlisstributeArticle(Enlight_Event_EventArgs $eventArgs)
-    {
-        $currentTime = new DateTime();
-
-        /** @var \Shopware\CustomModels\Blisstribute\BlisstributeArticleType $entity */
-        $entity = $eventArgs->get('entity');
-        $entity->setCreatedAt($currentTime)
-            ->setModifiedAt($currentTime);
-    }
-
-    /**
-     * article type event fired before db update
-     *
-     * @param Enlight_Event_EventArgs $args
-     *
-     * @return void
-     */
-    public function preUpdateBlisstributeArticle(Enlight_Event_EventArgs $args)
-    {
-        /** @var \Shopware\CustomModels\Blisstribute\BlisstributeArticle $entity */
-        $entity = $args->get('entity');
-        $entity->setModifiedAt(new DateTime());
-    }   
-
-    /**
-     * article type event fired before create entity
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return void
-     */
-    public function prePersistBlisstributeArticleType(Enlight_Event_EventArgs $eventArgs)
-    {
-        $currentTime = new DateTime();
-
-        /** @var \Shopware\CustomModels\Blisstribute\BlisstributeArticleType $entity */
-        $entity = $eventArgs->get('entity');
-        $entity->setCreatedAt($currentTime)
-            ->setModifiedAt($currentTime);
-    }
-
-    /**
-     * article type event fired before update
-     *
-     * @param Enlight_Event_EventArgs $args
-     *
-     * @return void
-     */
-    public function preUpdateBlisstributeArticleType(Enlight_Event_EventArgs $args)
-    {
-        $articleIdCollection = array();
-
-        $modelManager = $this->get('models');
-
-        /* @var Shopware\CustomModels\Blisstribute\BlisstributeArticleType $entity */
-        $entity = $args->get('entity');
-        $entity->setModifiedAt(new DateTime());
-
-        /* @var Shopware\CustomModels\Blisstribute\BlisstributeArticleRepository $articleRepository */
-        $articleRepository = $modelManager->getRepository('Shopware\CustomModels\Blisstribute\BlisstributeArticle');
-
-        $groupEntity = $entity->getFilter();
-        foreach ($groupEntity->getArticles() as $article) {
-            if (!in_array($article->getId(), $articleIdCollection)) {
-                $articleIdCollection[] = $article->getId();
-            }
-        }
-
-        $articleCollection = $articleRepository->fetchByArticleIdList($articleIdCollection);
-        foreach ($articleCollection as $currentArticle) {
-            $currentArticle->setModifiedAt(new DateTime());
-            $currentArticle->setTriggerSync(true);
-            $currentArticle->setTries(0);
-
-            $modelManager->persist($currentArticle);
-        }
-
-        $modelManager->flush();
-    }
-
-    /**
-     * @param \Enlight_Event_EventArgs $args
-     *
-     * @return void
-     */
-    public function prePersistShop(Enlight_Event_EventArgs $args)
-    {
-        $modelManager = $this->get('models');
-
-        /** @var \Shopware\Models\Shop\Shop $shop */
-        $shop = $args->get('entity');
-
-        $blisstributeShop = new \Shopware\CustomModels\Blisstribute\BlisstributeShop();
-        $blisstributeShop->setShop($shop)->setAdvertisingMediumCode('');
-
-        $modelManager->persist($blisstributeShop);
-        $modelManager->flush();
-    }
-
-    /**
-     * @param \Enlight_Event_EventArgs $args
-     *
-     * @return void
-     */
-    public function preRemoveShop(Enlight_Event_EventArgs $args)
-    {
-        $modelManager = $this->get('models');
-
-        /** @var Shopware\Models\Shop\ $blisstributeShop */
-        $shop = $args->get('entity');
-
-        $repository = $this->get('models')->getRepository('Shopware\CustomModels\Blisstribute\BlisstributeShop');
-        $blisstributeShop = $repository->findOneByShop($shop->getId());
-        if ($blisstributeShop === null) {
-            return;
-        }
-
-        $modelManager->remove($blisstributeShop);
-        $modelManager->flush();
-    }
-
-    /**
-     * @param \Enlight_Event_EventArgs $args
-     *
-     * @return void
-     */
-    public function prePersistVoucher(Enlight_Event_EventArgs $args)
-    {
-        $modelManager = $this->get('models');
-
-        /** @var \Shopware\Models\Voucher\Voucher $voucher */
-        $voucher = $args->get('entity');
-
-        $blisstributeCoupon = new \Shopware\CustomModels\Blisstribute\BlisstributeCoupon();
-        $blisstributeCoupon->setVoucher($voucher)->setIsMoneyVoucher(false);
-
-        $modelManager->persist($blisstributeCoupon);
-        $modelManager->flush();
-    }
-
-    /**
-     * Post Persist Voucher Attribute
-     * Automating Money Voucher (de: Wertgutschein) automation
-     *
-     * @param Enlight_Event_EventArgs $args
-     */
-    public function prePersistVoucherAttribute(Enlight_Event_EventArgs $args)
-    {
-        $modelManager = $this->get('models');
-
-        /** @var \Shopware\Models\Attribute\Voucher $attribute */
-        $attribute = $args->get('entity');
-
-        if (!is_null($attribute)) {            
-            $plugin = $this->get('models')->getRepository('Shopware\Models\Plugin\Plugin')->findOneBy(array(
-                'name' => 'NetiEasyCoupon',
-                'active' => true
-            ));
-            
-            if ($plugin) {
-                $blisstributeCoupon = $modelManager->getRepository('Shopware\CustomModels\Blisstribute\BlisstributeCoupon')->findOneBy(array(
-                    'voucher' => $attribute->getVoucher()
-                ));
-
-                $blisstributeCoupon->setIsMoneyVoucher($netiEasyCoupon);
-
-                $modelManager->persist($blisstributeCoupon);
-                $modelManager->flush();
-            }
-        }
-    }
-
-    /**
-     * @param \Enlight_Event_EventArgs $args
-     *
-     * @return void
-     */
-    public function preRemoveVoucher(Enlight_Event_EventArgs $args)
-    {
-        $modelManager = $this->get('models');
-
-        /** @var \Shopware\Models\Voucher\Voucher $voucher */
-        $voucher = $args->get('entity');
-
-        $repository = $this->get('models')->getRepository('Shopware\CustomModels\Blisstribute\BlisstributeCoupon');
-        $blisstributeCoupon = $repository->findByCoupon($voucher->getId());
-        if ($blisstributeCoupon === null) {
-            return;
-        }
-
-        $modelManager->remove($blisstributeCoupon);
-        $modelManager->flush();
-    }
-
-    /**
-     * @param \Enlight_Event_EventArgs $args
-     *
-     * @return void
-     */
-    public function preRemoveOrder(Enlight_Event_EventArgs $args)
-    {
-        $modelManager = $this->get('models');
-
-        /** @var \Shopware\Models\Order\Order $order */
-        $order = $args->get('entity');
-
-        $repository = $this->get('models')->getRepository('Shopware\CustomModels\Blisstribute\BlisstributeOrder');
-        $blisstributeOrder = $repository->findByOrder($order);
-        if ($blisstributeOrder === null) {
-            return;
-        }
-
-        $modelManager->remove($blisstributeOrder);
-        $modelManager->flush();
-    }
-
-    /**
-     * transfer all necessary article - event triggered by cron job
-     *
-     * @param Enlight_Event_EventArgs $args
-     *
-     * @return bool
-     *
-     * @deprecated removed in version 0.2.29
-     */
-    public function onBlisstributeArticleSync(Enlight_Event_EventArgs $args)
-    {
-        require_once __DIR__ . '/Components/Blisstribute/Article/Sync.php';
-
-        try {
-            $this->registerCustomModels();
-
-            $controller = new Shopware_Components_Blisstribute_Article_Sync($this->Config());
-            $controller->processBatchArticleSync();
-        } catch (Exception $e) {
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
-     * transfer all open orders to blisstribute - event triggered by cron job
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return bool
-     *
-     * @deprecated removed in version 0.2.30
-     */
-    public function onBlisstributeOrderExport(Enlight_Event_EventArgs $eventArgs)
-    {
-        require_once __DIR__ . '/Components/Blisstribute/Order/Sync.php';
-
-        try {
-            $this->registerCustomModels();
-
-            $orderSyncCtrl = new Shopware_Components_Blisstribute_Order_Sync($this->Config());
-            $orderSyncCtrl->processBatchOrderSync();
-        } catch (Exception $ex) {
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
-     * add order reference for export to blisstribute
-     *
-     * @param \Enlight_Event_EventArgs $eventArgs
-     *
-     * @return bool
-     */
-    public function onRegisterOrder(Enlight_Event_EventArgs $eventArgs)
-    {
-        $this->registerOrder($eventArgs);
-        return true;
-    }
-
-    /**
-     * export order to blisstribute
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return bool|string
-     */
-    public function onOrderFinished(Enlight_Event_EventArgs $eventArgs)
-    {
-        $blisstributeOrder = $this->registerOrder($eventArgs);
-        if ($blisstributeOrder === null || !$blisstributeOrder) {
-            $this->logInfo('blisstributeOrder is null! onOrderFinished failed!');
-            return false;
-        }
-
-        $order = $blisstributeOrder->getOrder();
-        if ($order === null || !$order) {
-            $this->logInfo('order is null! onOrderFinished failed!');
-            return false;
-        }
-
-        $this->logInfo('processing order ' . $order->getNumber());
-        if ($this->get('config')->get('googleAddressValidation')) {
-            /** @var Shopware_Components_Blisstribute_Order_GoogleAddressValidator $addressValidator */
-            $addressValidator = $this->get('blisstribute.google_address_validator');
-            $addressValidator->validateAddress($blisstributeOrder, $this->get('config'));
-        }
-
-        if ($blisstributeOrder->getStatus() == \Shopware\CustomModels\Blisstribute\BlisstributeOrder::EXPORT_STATUS_TRANSFERRED
-            || $blisstributeOrder->getStatus() == \Shopware\CustomModels\Blisstribute\BlisstributeOrder::EXPORT_STATUS_IN_TRANSFER
-        ) {
-            return true;
-        }
-
-        if (!$this->get('config')->get('blisstribute-auto-sync-order')) {
-            $this->logDebug('order sync cancelled due to disabled automatic sync');
-            return true;
-        }
-
-        try {
-            require_once __DIR__ . '/Components/Blisstribute/Order/Sync.php';
-            $orderSync = new Shopware_Components_Blisstribute_Order_Sync($this->Config());
-            $result = $orderSync->processSingleOrderSync($blisstributeOrder);
-        } catch (Exception $ex) {
-            return $ex->getMessage();
-        }
-
-        return $result;
-    }
-
-    /**
-     * register order to blisstribute order export
-     *
-     * @param Enlight_Event_EventArgs $eventArgs
-     *
-     * @return \Shopware\CustomModels\Blisstribute\BlisstributeOrder|null
-     */
-    protected function registerOrder(Enlight_Event_EventArgs $eventArgs)
-    {
-        $orderProxy = $eventArgs->get('subject');
-
-        $modelManager = $this->get('models');
-        $orderRepository = $modelManager->getRepository('Shopware\Models\Order\Order');
-
-        /** @var \Shopware\Models\Order\Order $order */
-        $order = $orderRepository->findOneBy(array('number' => $orderProxy->sOrderNumber));
-        if ($order === null) {
-            return null;
-        }
-
-        /** @var \Shopware\CustomModels\Blisstribute\BlisstributeOrderRepository $blisstributeOrderRepository */
-        $blisstributeOrderRepository = $modelManager->getRepository('Shopware\CustomModels\Blisstribute\BlisstributeOrder');
-        $blisstributeOrder = $blisstributeOrderRepository->findByOrder($order);
-        if ($blisstributeOrder === null) {
-            $status = \Shopware\CustomModels\Blisstribute\BlisstributeOrder::EXPORT_STATUS_CREATION_PENDING;
-
-            if (!$this->get('config')->get('blisstribute-auto-sync-order')) {
-                $status = \Shopware\CustomModels\Blisstribute\BlisstributeOrder::EXPORT_STATUS_NONE;
-            }
-        
-            $blisstributeOrder = new \Shopware\CustomModels\Blisstribute\BlisstributeOrder();
-            $blisstributeOrder->setLastCronAt(new DateTime())
-                ->setOrder($order)
-                ->setStatus($status)
-                ->setTries(0);
-
-            $modelManager->persist($blisstributeOrder);
-            $modelManager->flush();
-        }
-
-        return $blisstributeOrder;
     }
 
     /**
